requirements.txt
nats-py
ddecon
pyTelegramBotAPI
aiohttp
python-dotenv
nextcord
pydantic
schedule